Feature: Moved 'Dispatcher' to separated module.
